requirements.txt
obs-websocket-py
opencv-python
numpy
pillow

pywebview
customtkinter
darkdetect
python-dotenv
